removed all doubles from game.  This breaks the torpedoes, and sublight navigation.  Both those will be fixed.

+++ trunk/StarTrek_KG/Actors/Ship.cs
@@ -33,7 +33,7 @@ namespace StarTrek_KG.Actors
 
             public string Name { get; set; }
             public FactionName Faction { get; set; }
-            public double Energy { get; set; }
+            public int Energy { get; set; }
             public bool Destroyed { get; set; }
 
             ////todo: status of the battles will be kept in the ships LOG.  If you board a ship, you can read its log and see who it had a battle with.
